requirements.txt
langchain
langchain-google-genai
google-generativeai
pypdf
faiss-cpu
streamlit
python-dotenv
